requirements.txt
aiosqlite==0.17.0
anyio==3.5.0
ariadne==0.14.1
astroid==2.9.3
atomicwrites==1.4.0
attrs==21.4.0
beautifulsoup4==4.11.1
black==22.1.0
bs4==0.0.1
click==8.0.3
colorama==0.4.4
coverage==6.4
Flask==2.0.2
graphql-core==3.1.7
idna==3.3
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
lxml==4.8.0
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.1
pluggy==1.0.0
py==1.11.0
pylint==2.12.2
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
python-dotenv==0.20.0
sniffio==1.2.0
soupsieve==2.3.2.post1
starlette==0.17.1
toml==0.10.2
tomli==2.0.0
typing_extensions==4.0.1
Werkzeug==2.0.3
wrapt==1.13.3
